requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
gunicorn==22.0.0
flask-cors==4.0.0
limits==3.13.0
flask-limiter==3.7.0
openai==0.27.0
gTTS==2.3.2
pyttsx3==2.90
pydub==0.25.1 